PartyFinder namespaces, break out
